+++ app/models/unidom/geo/location.rb
@@ -4,6 +4,8 @@ class Unidom::Geo::Location < ActiveRecord::Base
 
   self.table_name = 'unidom_locations'
 
+  include Unidom::Common::Concerns::ModelExtension
+
   validates :latitude,  presence: true, numericality: { greater_than_or_equal_to: -90,  less_than_or_equal_to: 90  }
   validates :longitude, presence: true, numericality: { greater_than_or_equal_to: -180, less_than_or_equal_to: 180 }
 
@@ -17,9 +19,9 @@ class Unidom::Geo::Location < ActiveRecord::Base
 
   belongs_to :region, polymorphic: true
 
+  has_many :locatings, class_name: 'Unidom::Geo::Locating'
+
   scope :region_is,         ->(region)         { where region:         region         }
   scope :postal_address_is, ->(postal_address) { where postal_address: postal_address }
-
-  include Unidom::Common::Concerns::ModelExtension
 
 end