+++ CHANGELOG.md
@@ -7,3 +7,9 @@
 ## v0.2
 1. Improved the Location model to include the Model Extension concern
 2. Improved the Locating model to include the Model Extension concern
+
+## v0.3
+1. Improved the Location model to add the has_may :locatings macro
+2. Improved the Locating model for the class name
+3. Improved the Locating model to add the #locate! method
+4. Improved the Ruby Gem Specification to depend on [unidom-common](https://github.com/topbitdu/unidom-common) v0.9